start migrating to cmake

+++ id.c
@@ -80,5 +80,5 @@ void frudp_stuff_guid(frudp_guid_t *guid,
 void frudp_print_guid(frudp_guid_t *guid)
 {
   frudp_print_guid_prefix(&guid->guid_prefix);
-  printf("%08x", guid->entity_id.u);
+  printf("%08x", (unsigned)guid->entity_id.u);
 }
